fix(hcl2cdk): move coerceType call into new method that handles ast conversion
prepares us to do the type coercion for other layers, too. E.g. for arguments of function calls
