+++ packages/@cdktf/hcl2cdk/lib/expressions.ts
@@ -3,6 +3,7 @@
 import * as t from "@babel/types";
 import template from "@babel/template";
 import reservedWords from "reserved-words";
+import { toSnakeCase } from "codemaker";
 import { camelCase, logger, pascalCase } from "./utils";
 import { TerraformResourceBlock, Scope } from "./types";
 import {
@@ -16,6 +17,9 @@ import {
   Range,
 } from "@cdktf/hcl2json";
 import { getFullProviderName } from "./provider";
+import { coerceType } from "./coerceType";
+import { getDesiredType, getTypeAtPath } from "./terraformSchema";
+import { AttributeType } from "@cdktf/provider-generator";
 
 export type Reference = {
   start: number;
@@ -461,8 +465,12 @@ export const terraformObjectToTsAst = async (
 
   const ast = referencesToAst(scope, input, references, scopedIds);
 
-  const wrapInArray = isListExpression(input);
-  return wrapInArray ? t.arrayExpression([ast]) : ast;
+  return coerceType(
+    scope,
+    ast,
+    findTypeOfReference(scope, ast, references),
+    getDesiredType(scope, path)
+  );
 };
 
 export function referencesToAst( // TODO: include this into ast handling (everything comes from one single big JSON -> TS AST)
@@ -613,4 +621,84 @@ export async function findUsedReferences(
       )
     )
   ).flat();
+}
+
+/*
+ * Transforms a babel AST into a list of string accessors
+ * e.g. foo.bar.baz -> ["foo", "bar", "baz"]
+ */
+function destructureAst(ast: t.Expression): string[] | undefined {
+  switch (ast.type) {
+    case "Identifier":
+      return [ast.name];
+    case "MemberExpression":
+      const object = destructureAst(ast.object);
+      const property = destructureAst(ast.property as t.Expression);
+      if (object && property) {
+        return [...object, ...property];
+      } else {
+        return undefined;
+      }
+    default:
+      return undefined;
+  }
+}
+
+function findTypeOfReference(
+  scope: Scope,
+  ast: t.Expression,
+  references: Reference[]
+): AttributeType {
+  const isReferenceWithoutTemplateString =
+    ast.type === "MemberExpression" &&
+    ast.object.type === "Identifier" &&
+    references.length === 1;
+  // If we only have one reference this is a
+  if (isReferenceWithoutTemplateString) {
+    if (references[0].isVariable) {
+      return "dynamic";
+    }
+
+    const destructuredAst = destructureAst(ast);
+    if (!destructuredAst) {
+      logger.debug(
+        `Could not destructure ast: ${JSON.stringify(ast, null, 2)}`
+      );
+      return "dynamic";
+    }
+
+    const [astVariableName, ...attributes] = destructuredAst;
+    const variable = Object.values(scope.variables).find(
+      (x) => x.variableName === astVariableName
+    );
+
+    if (!variable) {
+      logger.debug(
+        `Could not find variable ${astVariableName} given scope: ${JSON.stringify(
+          scope.variables,
+          null,
+          2
+        )}`
+      );
+      // We don't know, this should not happen, but if it does we assume the worst case and make it dynamic
+      return "dynamic";
+    }
+
+    const { resource: resourceType } = variable;
+    const [provider, ...resourceNameFragments] = resourceType.split("_");
+    const tfResourcePath = `${provider}.${resourceNameFragments.join(
+      "_"
+    )}.${attributes.map((x) => toSnakeCase(x)).join(".")}`;
+    const type = getTypeAtPath(scope.providerSchema, tfResourcePath);
+
+    // If this is an attribute type we can return it
+    if (typeof type === "string" || Array.isArray(type)) {
+      return type;
+    }
+
+    // Either nothing is found or it's a block type
+    return "dynamic";
+  }
+
+  return "string";
 }
