+++ packages/@cdktf/hcl2cdk/test/terraformExpressionToAst.test.ts
@@ -7,6 +7,7 @@ const mockScope: Scope = {
   providerGenerator: {},
   providerSchema: {},
   variables: {},
+  hasTokenBasedTypeCoercion: false,
 };
 
 const mockPath = "";
